<v0.12.0> #163 - Update Core Types (int -> size_t / ptrdiff_t)
- Changed core types to use size_t / ptrdiff_t.
- Methods that previously returned -1 for "none" now return std::string::npos.
- Qt render types remain as int because Qt expects int values.
- Updated Windows CMake instructions in README.md for different configurations.
- Fixed tab command type issues and a bug where the tab was not always applied.
- Fixed MinGW errors in tests
  - Using variadic functions with size_t caused issues with va_args and type casting.
  - The va_args macro can behave inconsistently on Windows.
  - Resolved by switching to a C++ initializer list.

+++ edbee-lib/edbee/commands/debugcommand.cpp
@@ -24,39 +24,41 @@ DebugCommand::DebugCommand(DebugCommandType command)
 void DebugCommand::execute(TextEditorController* controller)
 {
     switch( command_ ) {
-        case DumpScopes: dumpScopes( controller ); break;
-        case RebuildScopes: rebuildScopes( controller ); break;
-        case DumpUndoStack: dumpUndoStack( controller ); break;
-        case DumpCharacterCodes: dumpCharacterCodes( controller ); break;
+        case DumpScopes: dumpScopes(controller); break;
+        case RebuildScopes: rebuildScopes(controller); break;
+        case DumpUndoStack: dumpUndoStack(controller); break;
+        case DumpCharacterCodes: dumpCharacterCodes(controller); break;
     }
-
 }
+
 
 QString DebugCommand::toString()
 {
     return "DebugCommand";
 }
 
+
 bool DebugCommand::readonly()
 {
     return true;
 }
 
+
 /// This method dumps the scopes
-void DebugCommand::dumpScopes( TextEditorController* controller )
+void DebugCommand::dumpScopes(TextEditorController* controller)
 {
     TextDocumentScopes* scopes = controller->textDocument()->scopes();
     QString dump = "\nMultiLineScopes:---\n";
-    dump.append( QString(scopes->toString()).replace("|","\n") );
+    dump.append(QString(scopes->toString()).replace("|","\n"));
 
     dump.append("\nPer Line:---\n");
-    for( int i=0; i<scopes->scopedLineCount(); ++i ) {
+    for (size_t i = 0; i < scopes->scopedLineCount(); ++i) {
         ScopedTextRangeList* range = scopes->scopedRangesAtLine(i);
-        dump.append( QStringLiteral("%1: %2\n").arg(i).arg( range->toString() ) );
+        dump.append(QStringLiteral("%1: %2\n").arg(i).arg( range->toString()));
     }
-
     qlog_info() << dump;
 }
+
 
 /// rebuilds all scopes
 void DebugCommand::rebuildScopes(TextEditorController* controller)
@@ -67,7 +69,7 @@ void DebugCommand::rebuildScopes(TextEditorController* controller)
 
     /// lex the complete document
     scopes->removeScopesAfterOffset(0);
-    lexer->lexRange( 0, document->length() );
+    lexer->lexRange(0, document->length());
 }
 
 /// dumps the undostack
@@ -78,9 +80,9 @@ void DebugCommand::dumpUndoStack(TextEditorController* controller)
 
 
 /// dumps the character codes (around the FIRST caret)
-void DebugCommand::dumpCharacterCodes(TextEditorController *controller)
+void DebugCommand::dumpCharacterCodes(TextEditorController* controller)
 {
-    static int count = 3;       // 2 chars before caret and 2 chars after caret
+    static size_t count = 3;       // 2 chars before caret and 2 chars after caret
 
     // get the current range
     TextSelection* sel = controller->textSelection();
@@ -89,20 +91,18 @@ void DebugCommand::dumpCharacterCodes(TextEditorController *controller)
     QString line;
 
     // iterate over the characters
-    int start = qMax(range.caret() - count, 0);
-    int end = qMin(range.caret() + count, controller->textDocument()->length()-1 );
-    for( int i=start; i<=end; ++i ) {
+    size_t start = qMax(range.caret() - count, static_cast<size_t>(0));
+    size_t end = qMin(range.caret() + count, controller->textDocument()->length()-1 );
+    for (size_t i = start; i <= end; ++i) {
         QChar c = controller->textDocument()->charAt(i);
-        if( i == range.caret() ) {
+        if (i == range.caret()) {
             line.append("| ");
         }
 
-        line.append( QString::number((int)c.unicode(), 16) );
+        line.append(QString::number((int)c.unicode(), 16));
         line.append(" ");
     }
-
     qlog_info() << "charcodes: " <<  line;
 }
-
 
 } // edbee
